requirements.txt
Flask~=2.1.2
torch~=1.12.0
scikit-image~=0.19.3
numpy~=1.23.1
matplotlib~=3.5.2
celery~=4.4.6
Pillow~=9.2.0
gunicorn~=20.1.0